Move code to shared functions
