+++ packages/storage/test/unit/task.test.ts
@@ -19,175 +19,13 @@ import { FbsBlob } from '../../src/implementation/blob';
 import { Location } from '../../src/implementation/location';
 import { Unsubscribe } from '../../src/implementation/observer';
 import { TaskEvent, TaskState } from '../../src/implementation/taskenums';
-import { Headers } from '../../src/implementation/xhrio';
 import { Reference } from '../../src/reference';
-import { StorageService } from '../../src/service';
 import { UploadTask } from '../../src/task';
-import { makePool, emptyAuthProvider } from './testshared';
-import { StringHeaders, TestingXhrIo } from './xhrio';
-import { FirebaseApp } from '@firebase/app-types';
+import { fakeServerHandler, storageServiceWithHandler } from './testshared';
 
 const testLocation = new Location('bucket', 'object');
 const smallBlob = new FbsBlob(new Blob(['a']));
 const bigBlob = new FbsBlob(new Blob([new ArrayBuffer(1024 * 1024)]));
-
-const fakeMetadata = '{ "downloadTokens": "a,b" }';
-
-interface Response {
-  status: number;
-  body: string;
-  headers: StringHeaders;
-}
-type RequestHandler = (
-  url: string,
-  method: string,
-  body?: ArrayBufferView | Blob | string | null,
-  headers?: Headers
-) => Response;
-
-function storageServiceWithHandler(handler: RequestHandler): StorageService {
-  function newSend(
-    xhrio: TestingXhrIo,
-    url: string,
-    method: string,
-    body?: ArrayBufferView | Blob | string | null,
-    headers?: Headers
-  ): void {
-    const response = handler(url, method, body, headers);
-    xhrio.simulateResponse(response.status, response.body, response.headers);
-  }
-
-  return new StorageService(
-    {} as FirebaseApp,
-    emptyAuthProvider,
-    makePool(newSend)
-  );
-}
-
-function fakeServerHandler(): RequestHandler {
-  const stats: {
-    [num: number]: {
-      currentSize: number;
-      finalSize: number;
-    };
-  } = {};
-
-  let nextId: number = 0;
-
-  function statusHeaders(
-    status: string,
-    existing?: StringHeaders
-  ): StringHeaders {
-    if (existing) {
-      existing['X-Goog-Upload-Status'] = status;
-      return existing;
-    } else {
-      return { 'X-Goog-Upload-Status': status };
-    }
-  }
-
-  function handler(
-    url: string,
-    method: string,
-    content?: ArrayBufferView | Blob | string | null,
-    headers?: Headers
-  ): Response {
-    method = method || 'GET';
-    content = content || '';
-    headers = headers || {};
-
-    if (headers['X-Goog-Upload-Protocol'] === 'multipart') {
-      return {
-        status: 200,
-        body: fakeMetadata,
-        headers: statusHeaders('final')
-      };
-    }
-
-    const contentLength =
-      (content as Blob).size || (content as string).length || 0;
-
-    if (headers['X-Goog-Upload-Protocol'] === 'resumable') {
-      const thisId = nextId;
-      nextId++;
-      stats[thisId] = {
-        currentSize: 0,
-        finalSize: +headers['X-Goog-Upload-Header-Content-Length']
-      };
-
-      return {
-        status: 200,
-        body: '',
-        headers: statusHeaders('active', {
-          'X-Goog-Upload-URL': 'http://example.com?' + thisId
-        })
-      };
-    }
-
-    const matches = url.match(/^http:\/\/example\.com\?([0-9]+)$/);
-    if (matches === null) {
-      return { status: 400, body: '', headers: {} };
-    }
-
-    const id = +matches[1];
-    if (!stats[id]) {
-      return { status: 400, body: 'Invalid upload id', headers: {} };
-    }
-
-    if (headers['X-Goog-Upload-Command'] === 'query') {
-      return {
-        status: 200,
-        body: '',
-        headers: statusHeaders('active', {
-          'X-Goog-Upload-Size-Received': stats[id].currentSize.toString()
-        })
-      };
-    }
-
-    const commands = (headers['X-Goog-Upload-Command'] as string)
-      .split(',')
-      .map(str => {
-        return str.trim();
-      });
-    const isUpload = commands.indexOf('upload') !== -1;
-    const isFinalize = commands.indexOf('finalize') !== -1;
-    const stat = stats[id];
-
-    if (isUpload) {
-      const offset = +headers['X-Goog-Upload-Offset'];
-      if (offset !== stat.currentSize) {
-        return { status: 400, body: 'Uploading at wrong offset', headers: {} };
-      }
-
-      stat.currentSize += contentLength;
-      if (stat.currentSize > stat.finalSize) {
-        return { status: 400, body: 'Too many bytes', headers: {} };
-      } else if (!isFinalize) {
-        return { status: 200, body: '', headers: statusHeaders('active') };
-      }
-    }
-
-    if (isFinalize) {
-      const isDone = stat.currentSize === stat.finalSize;
-      if (isDone) {
-        return {
-          status: 200,
-          body: fakeMetadata,
-          headers: statusHeaders('final')
-        };
-      } else {
-        return {
-          status: 400,
-          body: 'finalize without the right # of bytes',
-          headers: {}
-        };
-      }
-    }
-
-    return { status: 400, body: '', headers: {} };
-  }
-  return handler;
-}
 
 describe('Firebase Storage > Upload Task', () => {
   it('Works for a small upload w/ an observer', done => {
